Create README.md
The game of Mafia is about convincing others, about being able to lie believably, and about being able to figure out if other people are lying. It will reveal a lot about players to each other, and works best if the players know a little bit about each other first. It is tremendous fun if the players have a sense of humor.