How to Play Mafia The game of Mafia is about convincing others, about being able to lie believably, and about being able to figure out if other people are lying. It will reveal a lot about players to each other, and works best if the players know a little bit about each other first. It is tremendous fun if the players have a sense of humor.
How to Play Mafia
The game of Mafia is about convincing others, about being able to lie
believably, and about being able to figure out if other people are
lying. It will reveal a lot about players to each other, and works best
if the players know a little bit about each other first. It is
tremendous fun if the players have a sense of humor.
The setting is a once peaceful town which has been invaded by the dark
forces of corruption - the Mafia.
The Mafia's sole purpose is to murder the productive citizens in their
bed. They have free reign of the town at night, but by day they appear
to be normal citizens.
As a response to these terrible events, the townspeople have hired an
Inspector, whose job is to remain disguised while attempting to root our
the members of the Mafia.
________________________________________
Game Rules
Game Setup
The game of Mafia is played by an odd number of people. Good numbers are
9, 11, or 13. Players are either townspeople or Mafia. In addition, one
of the townspeople is the Inspector. Who plays what is determined at
random. One method for determining who is who is to use cards: black
cards for the mafia, red cards for townspeople, and the king of hearts
for the inspector. Players may not show their cards to others.
The number of Mafia and townspeople can vary, depending on which side is
winning more often:
Players Mafia Leans Toward
9 3 Even
9 2 Townspeople
11 3 Even
11 4 Mafia
13 3 Townspeople
13 4 Even
13 5 Mafia
Starting with 3 Mafia is generally recommended.
Order of Play
The game takes place over the duration of several "days" and "nights".
Play begins at the start of the first "day", and proceeds as follows:
1. Everyone discusses who might be in the Mafia. Eventually, a vote is
called and the (un)lucky candidate is run out of town (or lynched, or
tarred and feathered, or... You get the idea). As they are run out of
town (of whatever), the player gets to make one final statement.
The first person thus removed from play becomes the Moderator
2. Night falls. Everyone goes to sleep (by closing their eyes).
3. The Moderator tells the members of the Mafia to open their eyes. The
members of the Mafia then select someone they will murder, by pointing
at people until they all agree on one. The Mafia then close their eyes.
4. The Moderator tells the Inspector to open his eyes. This happens even
if the Inspector has been killed, or run out of town. The Inspector then
points at a suspect. The Moderator nods "yes" if that person in the
Mafia, or "no" if they are innocent. The Inspector then closes his eyes.
5. The Moderator announces the coming of the morning, and the name of
the luckless victim. Embelishments on the method of death are fun here
(Ooh, Jim was found garotted in his Barcalounger!).
6. The townspeople, realizing that their previous council did not root
out all the evil, return to step one and play proceeds from there.
Winning Conditions
The game is over when:
• All of the Mafia have been eliminated. The townspeople win
• There are an equal number of townspeople and Mafia at the end of a
"day". The Mafia wins
Play Hints
Calling a Vote
During the day, the townspeople get to vote on who to run out of town.
At any point during the discussion, someone can call a vote by saying "I
vote So-and-So is a Mafia and should be run out of town." If more than
half the living players agree, then the elected person is
unceremoniously removed from the game. There should always be an odd
number of players during the day, so a vote will not end in a deadlock.
Voting and debated may become heated. That's where the fun lies!
The First Day
During the first day, members of the Mafia have no information about
who's in the Mafia or not. The person to get voted out is pretty random,
so it's up to everyone to make a good, convincing first statement that
will result in them not being voted out.
Once the first night falls, though, the Mafia will be able to act in a
much mure unified way.
The Inspector
The Inspector has no way to prove their role, since players are not
allowed to show their cards to one another to prove they are what they
say there are. Admitting one is the Inspector can result in a swift
death in the middle of the night, unless only one member of the Mafia
was left alive.